To python3

+++ Automation/psi_scaleway.py
@@ -162,7 +162,7 @@ class PsiScaleway:
             vol_res = self.client.query().volumes(scaleway['volumes']['0']['id']).delete()
 
             # Delete IPs
-            if scaleway['public_ip']['dynamic'] == False:
+            if scaleway['public_ip'] != None and scaleway['public_ip']['dynamic'] == False:
                 ip_res = self.client.query().ips(scaleway['public_ip']['id']).delete()
         except slexc.HttpClientError as exc:
             print(json.dumps(exc.response.json(), indent=2))
@@ -376,4 +376,4 @@ def launch_new_server(scaleway_account, is_TCS, plugins, multi_ip=False):
             datacenter_name, region, None, None)
 
 if __name__ == '__main__':
-    print launch_new_server()
+    print(launch_new_server())
